+++ tests/functional/m/missing/missing_yield_doc_required.py
@@ -1,69 +1,7 @@
 """Tests for missing-yield-doc and missing-yield-type-doc with accept-no-yields-doc = no"""
 # pylint: disable=missing-function-docstring, unused-argument, function-redefined
-import mymodule  # [import-error]
 
 
 # Test missing docstring
 def my_func(self):  # [missing-yield-doc, missing-yield-type-doc]
-    yield False
-
-
-# Test Sphinx docstring
-def my_func(self):
-    """This is a docstring.
-
-    :return: Always False
-    :rtype: bool
-    """
-    yield False
-
-
-def my_func(self):
-    """This is a docstring.
-
-    :returns: An object
-    :rtype: :class:`mymodule.Class`
-    """
-    yield mymodule.Class()
-
-
-def my_func(self):
-    """This is a docstring.
-
-    :returns: An object
-    :rtype: list(:class:`mymodule.Class`)
-    """
-    yield [mymodule.Class()]
-
-
-def my_func(self):  # [missing-yield-doc]
-    """This is a docstring.
-
-    :rtype: list(:class:`mymodule.Class`)
-    """
-    yield [mymodule.Class()]
-
-
-def my_func(self):  # [missing-yield-type-doc]
-    """This is a docstring.
-
-    :returns: Always False
-    """
-    yield False
-
-
-def my_func(self):  # [missing-yield-doc]
-    """This is a docstring.
-
-    :rtype: bool
-    """
-    yield False
-
-
-def my_func(self, doc_type):  # [missing-yield-doc, missing-yield-type-doc]
-    """This is a docstring.
-
-    :param doc_type: Sphinx
-    :type doc_type: str
-    """
     yield False
