+++ tests/functional/m/missing/missing_yield_doc_required_Sphinx.txt
@@ -1,0 +1,5 @@
+missing-yield-doc:35:0:40:28:my_func:Missing yield documentation:UNDEFINED
+missing-yield-type-doc:43:0:48:15:my_func:Missing yield type documentation:UNDEFINED
+missing-yield-doc:51:0:56:15:my_func:Missing yield documentation:UNDEFINED
+missing-yield-doc:59:0:65:15:my_func:Missing yield documentation:UNDEFINED
+missing-yield-type-doc:59:0:65:15:my_func:Missing yield type documentation:UNDEFINED
